+++ bcbio/distributed/ipythontasks.py
@@ -2,7 +2,10 @@
 """
 import contextlib
 
-from IPython.parallel import require
+try:
+    from ipyparallel import require
+except ImportError:
+    from IPython.parallel import require
 
 from bcbio import heterogeneity, chipseq, structural, upload
 from bcbio.bam import callable
@@ -10,6 +13,8 @@ from bcbio.rnaseq import sailfish
 from bcbio.distributed import ipython
 from bcbio.ngsalign import alignprep
 from bcbio import rnaseq
+from bcbio.srna import sample as srna
+from bcbio.srna import group as seqcluster
 from bcbio.pipeline import (archive, config_utils, disambiguate, sample,
                             qcsummary, shared, variation, run_info, rnaseq)
 from bcbio.provenance import system
@@ -72,6 +77,36 @@ def trim_sample(*args):
     args = ipython.unzip_args(args)
     with _setup_logging(args) as config:
         return ipython.zip_args(apply(sample.trim_sample, *args))
+
+@require(srna)
+def trim_srna_sample(*args):
+    args = ipython.unzip_args(args)
+    with _setup_logging(args) as config:
+        return ipython.zip_args(apply(srna.trim_srna_sample, *args))
+
+@require(srna)
+def seqbuster(*args):
+    args = ipython.unzip_args(args)
+    with _setup_logging(args) as config:
+        return ipython.zip_args(apply(srna.mirbase, *args))
+
+@require(seqcluster)
+def seqcluster_prepare(*args):
+    args = ipython.unzip_args(args)
+    with _setup_logging(args) as config:
+        return ipython.zip_args(apply(seqcluster.run_prepare, *args))
+
+@require(seqcluster)
+def seqcluster_cluster(*args):
+    args = ipython.unzip_args(args)
+    with _setup_logging(args) as config:
+        return ipython.zip_args(apply(seqcluster.run_cluster, *args))
+
+@require(seqcluster)
+def srna_alignment(* args):
+    args = ipython.unzip_args(args)
+    with _setup_logging(args) as config:
+        return ipython.zip_args(apply(seqcluster.run_align, *args))
 
 @require(sailfish)
 def run_sailfish(*args):
@@ -150,6 +185,12 @@ def pipeline_summary(*args):
     args = ipython.unzip_args(args)
     with _setup_logging(args) as config:
         return ipython.zip_args(apply(qcsummary.pipeline_summary, *args))
+
+@require(qcsummary)
+def coverage_report(*args):
+    args = ipython.unzip_args(args)
+    with _setup_logging(args) as config:
+        return ipython.zip_args(apply(qcsummary.coverage_report, *args))
 
 @require(qcsummary)
 def qsignature_summary(*args):