+++ bcbio/provenance/programs.py
@@ -34,8 +34,7 @@ _cl_progs = [{"cmd": "bamtofastq", "name": "biobambam",
              {"cmd": "sambamba", "stdout_flag": "sambamba"},
              {"cmd": "qualimap", "args": "-h", "stdout_flag": "QualiMap"},
              {"cmd": "vcflib", "has_cl_version": False},
-             {"cmd": "featurecounts", "args": "-v", "stdout_flag": "featureCounts"},
-             {"cmd": "kraken", "args": "-v", "stdout_flag": "kraken"}]
+             {"cmd": "featurecounts", "args": "-v", "stdout_flag": "featureCounts"}]
 
 def _broad_versioner(type):
     def get_version(config):
@@ -72,7 +71,7 @@ def jar_versioner(program_name, jar_name):
             jar = jar.replace(to_remove, "")
         if jar.startswith(("-", ".")):
             jar = jar[1:]
-        if jar is "":
+        if not jar:
             logger.warn("Unable to determine version for program '{}' from jar file {}".format(
                 program_name, config_utils.get_jar(jar_name, pdir)))
         return jar