+++ bcbio/structural/lumpy.py
@@ -6,6 +6,7 @@ https://github.com/arq5x/lumpy-sv
 """
 import os
 import sys
+import shutil
 
 from bcbio import utils
 from bcbio.distributed.transaction import file_transaction, tx_tmpdir
@@ -38,8 +39,7 @@ def _run_lumpy(full_bams, sr_bams, disc_bams, work_dir, items):
                        "lumpyexpress -v -B {full_bams} -S {sr_bams} -D {disc_bams} "
                        "{exclude} -T {tmpdir} -o {tx_out_file}")
                 do.run(cmd.format(**locals()), "lumpyexpress", items[0])
-    sort_file = vcfutils.sort_by_ref(out_file, items[0])
-    return vcfutils.bgzip_and_index(sort_file, items[0]["config"]), sv_exclude_bed
+    return vcfutils.sort_by_ref(out_file, items[0]), sv_exclude_bed
 
 def _filter_by_support(in_file, data):
     """Filter call file based on supporting evidence, adding FILTER annotations to VCF.
@@ -54,13 +54,17 @@ def _filter_by_support(in_file, data):
     return vfilter.hard_w_expression(in_file, rc_filter, data, name="ReadCountSupport",
                                      limit_regions=None)
 
+def _sv_workdir(data):
+    return utils.safe_makedir(os.path.join(data["dirs"]["work"], "structural",
+                                           dd.get_sample_name(data), "lumpy"))
+
 def run(items):
     """Perform detection of structural variations with lumpy, using bwa-mem alignment.
     """
     if not all(utils.get_in(data, ("config", "algorithm", "aligner")) in ["bwa", False, None] for data in items):
         raise ValueError("Require bwa-mem alignment input for lumpy structural variation detection")
-    work_dir = utils.safe_makedir(os.path.join(items[0]["dirs"]["work"], "structural", items[0]["name"][-1],
-                                               "lumpy"))
+    paired = vcfutils.get_paired_bams([x["align_bam"] for x in items], items)
+    work_dir = _sv_workdir(paired.tumor_data if paired and paired.tumor_data else items[0])
     full_bams, sr_bams, disc_bams = [], [], []
     for data in items:
         dedup_bam, sr_bam, disc_bam = sshared.get_split_discordants(data, work_dir)
@@ -73,12 +77,31 @@ def run(items):
         if "sv" not in data:
             data["sv"] = []
         sample = dd.get_sample_name(data)
-        sample_vcf = _filter_by_support(vcfutils.select_sample(lumpy_vcf, sample,
-                                                               utils.append_stem(lumpy_vcf, "-%s" % sample),
-                                                               data["config"]),
-                                        data)
+        dedup_bam, sr_bam, _ = sshared.get_split_discordants(data, work_dir)
+        sample_vcf = vcfutils.select_sample(lumpy_vcf, sample,
+                                            utils.append_stem(lumpy_vcf, "-%s" % sample),
+                                            data["config"])
+        gt_vcf = _run_svtyper(sample_vcf, dedup_bam, sr_bam, data)
+        filter_vcf = _filter_by_support(gt_vcf, data)
         data["sv"].append({"variantcaller": "lumpy",
-                           "vrn_file": sample_vcf,
+                           "vrn_file": filter_vcf,
                            "exclude_file": exclude_file})
         out.append(data)
     return out
+
+def _run_svtyper(in_file, full_bam, sr_bam, data):
+    """Genotype structural variant calls with SVtyper.
+    """
+    out_file = "%s-wgts.vcf.gz" % utils.splitext_plus(in_file)[0]
+    if not utils.file_uptodate(out_file, in_file):
+        with file_transaction(data, out_file) as tx_out_file:
+            if not vcfutils.vcf_has_variants(in_file):
+                shutil.copy(in_file, out_file)
+            else:
+                python = sys.executable
+                svtyper = os.path.join(os.path.dirname(sys.executable), "svtyper")
+                cmd = ("gunzip -c {in_file} | "
+                       "{python} {svtyper} -B {full_bam} -S {sr_bam} | "
+                       "bgzip -c > {tx_out_file}")
+                do.run(cmd.format(**locals()), "SV genotyping with svtyper")
+    return vcfutils.sort_by_ref(out_file, data)